Make the translation of CleanFlight to Rust easier, by removing casted, nested struct constants.
It appears that the current version of C2Rust has trouble converting
certain nested C struct forms, generating a "convert vector not
supported" when it happens. This series of changes removes the nesting
that was causing those errors in some builds of CleanFlight; there may
be others that still need changing in a similar manner. A good TODO, if
we ever want to push this back upstream, would be to change the variable
names to more closely match CleanFlight patterns.